# packages/contracts/openapi/source-imports.openapi.json
{
  "components": {
    "schemas": {
      "FileValidateRequest": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "maxLength": 140000000,
            "minLength": 1,
            "title": "Content Base64",
            "type": "string"
          },
          "media_type": {
            "enum": [
              "text/csv",
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
            ],
            "title": "Media Type",
            "type": "string"
          }
        },
        "required": [
          "media_type",
          "content_base64"
        ],
        "title": "FileValidateRequest",
        "type": "object"
      },
      "FileValidateResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted_row_count": {
            "minimum": 0.0,
            "title": "Accepted Row Count",
            "type": "integer"
          },
          "content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Content Hash",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "rejected_rows": {
            "items": {
              "$ref": "#/components/schemas/RejectedRowResponse"
            },
            "title": "Rejected Rows",
            "type": "array"
          },
          "schema_fingerprint": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Schema Fingerprint",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "template_version": {
            "title": "Template Version",
            "type": "integer"
          }
        },
        "required": [
          "template_id",
          "template_version",
          "media_type",
          "content_hash",
          "schema_fingerprint",
          "accepted_row_count",
          "rejected_rows"
        ],
        "title": "FileValidateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RejectedRowResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "row_number": {
            "minimum": 1.0,
            "title": "Row Number",
            "type": "integer"
          }
        },
        "required": [
          "row_number",
          "code"
        ],
        "title": "RejectedRowResponse",
        "type": "object"
      },
      "TemplateColumnRequest": {
        "additionalProperties": false,
        "properties": {
          "data_type": {
            "enum": [
              "string",
              "integer",
              "decimal",
              "date",
              "datetime",
              "boolean"
            ],
            "title": "Data Type",
            "type": "string"
          },
          "field_role": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Field Role",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "source_header": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Source Header",
            "type": "string"
          }
        },
        "required": [
          "source_header",
          "field_role",
          "data_type"
        ],
        "title": "TemplateColumnRequest",
        "type": "object"
      },
      "TemplateColumnResponse": {
        "additionalProperties": false,
        "properties": {
          "data_type": {
            "enum": [
              "string",
              "integer",
              "decimal",
              "date",
              "datetime",
              "boolean"
            ],
            "title": "Data Type",
            "type": "string"
          },
          "field_role": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Field Role",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "source_header": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Source Header",
            "type": "string"
          }
        },
        "required": [
          "source_header",
          "field_role",
          "data_type"
        ],
        "title": "TemplateColumnResponse",
        "type": "object"
      },
      "TemplateCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "accepted_media_types": {
            "items": {
              "enum": [
                "text/csv",
                "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Accepted Media Types",
            "type": "array"
          },
          "allowed_sheet_names": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Allowed Sheet Names",
            "type": "array"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TemplateColumnRequest"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "date_format": {
            "default": "%Y-%m-%d",
            "maxLength": 80,
            "minLength": 2,
            "title": "Date Format",
            "type": "string"
          },
          "decimal_separator": {
            "default": ".",
            "enum": [
              ".",
              ","
            ],
            "title": "Decimal Separator",
            "type": "string"
          },
          "error_policy": {
            "default": "reject_file",
            "enum": [
              "reject_file",
              "reject_rows_with_report"
            ],
            "title": "Error Policy",
            "type": "string"
          },
          "file_size_limit": {
            "default": 10000000,
            "maximum": 100000000.0,
            "minimum": 1.0,
            "title": "File Size Limit",
            "type": "integer"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "required_sheet_names": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Required Sheet Names",
            "type": "array"
          },
          "row_limit": {
            "default": 100000,
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Row Limit",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "accepted_media_types",
          "columns"
        ],
        "title": "TemplateCreateRequest",
        "type": "object"
      },
      "TemplatePublishRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "TemplatePublishRequest",
        "type": "object"
      },
      "TemplateResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted_media_types": {
            "items": {
              "type": "string"
            },
            "title": "Accepted Media Types",
            "type": "array"
          },
          "allowed_sheet_names": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Sheet Names",
            "type": "array"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TemplateColumnResponse"
            },
            "title": "Columns",
            "type": "array"
          },
          "date_format": {
            "title": "Date Format",
            "type": "string"
          },
          "decimal_separator": {
            "enum": [
              ".",
              ","
            ],
            "title": "Decimal Separator",
            "type": "string"
          },
          "error_policy": {
            "enum": [
              "reject_file",
              "reject_rows_with_report"
            ],
            "title": "Error Policy",
            "type": "string"
          },
          "file_size_limit": {
            "title": "File Size Limit",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required_sheet_names": {
            "items": {
              "type": "string"
            },
            "title": "Required Sheet Names",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "row_limit": {
            "title": "Row Limit",
            "type": "integer"
          },
          "status": {
            "enum": [
              "draft",
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "template_id",
          "version",
          "revision",
          "name",
          "status",
          "accepted_media_types",
          "allowed_sheet_names",
          "required_sheet_names",
          "columns",
          "row_limit",
          "file_size_limit",
          "decimal_separator",
          "date_format",
          "error_policy"
        ],
        "title": "TemplateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Custometry File Import API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/templates": {
      "post": {
        "operationId": "create_file_import_template",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Template"
      }
    },
    "/templates/{template_id}/publish": {
      "post": {
        "operationId": "publish_file_import_template",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplatePublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Template"
      }
    },
    "/templates/{template_id}/validate": {
      "post": {
        "operationId": "validate_file_import",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate"
      }
    }
  }
}